# perf: make `ape test --help` faster