--- a/tests/functional/test_test.py
+++ b/tests/functional/test_test.py
@@ -1,8 +1,11 @@
+from pathlib import Path
+
 import pytest
 
 from ape.exceptions import ConfigError
 from ape.pytest.runners import PytestApeRunner
 from ape_test import ApeTestConfig
+from ape_test._watch import run_with_observer
 
 
 class TestApeTestConfig:
@@ -33,3 +36,30 @@ def test_connect_to_mainnet_by_default(mocker):
     )
     with pytest.raises(ConfigError, match=expected):
         runner._connect()
+
+
+def test_watch(mocker):
+    mock_event_handler = mocker.MagicMock()
+    event_handler_patch = mocker.patch("ape_test._watch._create_event_handler")
+    event_handler_patch.return_value = mock_event_handler
+
+    mock_observer = mocker.MagicMock()
+    observer_patch = mocker.patch("ape_test._watch._create_observer")
+    observer_patch.return_value = mock_observer
+
+    run_subprocess_patch = mocker.patch("ape_test._watch.run_subprocess")
+    run_main_loop_patch = mocker.patch("ape_test._watch._run_main_loop")
+    run_main_loop_patch.side_effect = SystemExit  # Avoid infinite loop.
+
+    # Only passing `-s` so we have an extra arg to test.
+    with pytest.raises(SystemExit):
+        run_with_observer((Path("contracts"),), 0.1, "-s")
+
+    # The observer started, then the main runner exits, and the observer stops + joins.
+    assert mock_observer.start.call_count == 1
+    assert mock_observer.stop.call_count == 1
+    assert mock_observer.join.call_count == 1
+
+    # NOTE: We had a bug once where the args it received were not strings.
+    #   (wasn't deconstructing), so this check is important.
+    run_subprocess_patch.assert_called_once_with(["ape", "test", "-s"])