### What I did

before:

```
ape test --help  1.59s user 0.20s system 99% cpu 1.803 total
```

now:

```
ape test --help  1.27s user 0.17s system 98% cpu 1.461 total
```

**note**: the time of `ape test --help` (e.g. `pytest --help`) largely depends on so many other things - conftests, plugins, etc.

### How I did it

move watch stuff out of cli directly and have it locally imported

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated